Migrate ProjectAppId on screen sdk
