+++ packages/server/src/sdk/app/screens/screens.ts
@@ -1,15 +1,46 @@
 import { context } from "@budibase/backend-core"
-import { Database, Screen } from "@budibase/types"
+import { Database, DocumentType, Screen, SEPARATOR } from "@budibase/types"
 import { getScreenParams } from "../../../db/utils"
+import sdk from "../.."
 
 export async function fetch(
   db: Database = context.getAppDB()
 ): Promise<Screen[]> {
-  return (
+  const screens = (
     await db.allDocs<Screen>(
       getScreenParams(null, {
         include_docs: true,
       })
     )
   ).rows.map(el => el.doc!)
+
+  if (screens.every(s => s.projectAppId)) {
+    return screens
+  }
+
+  await migrateToProjectApp(screens)
+  return fetch(db)
+}
+
+async function migrateToProjectApp(screens: Screen[]) {
+  const application = await sdk.applications.metadata.get()
+  if (screens.every(s => s.projectAppId)) {
+    return screens
+  }
+
+  // Forcing the _id to avoid concurrency issues
+  const createdProjectApp = await sdk.projectApps.update({
+    _id: `${DocumentType.PROJECT_APP}${SEPARATOR}default`,
+    name: application.name,
+  })
+
+  const db = context.getAppDB()
+  await db.bulkDocs(
+    screens.map<Screen>(s => ({
+      ...s,
+      projectAppId: createdProjectApp._id,
+    }))
+  )
+
+  return sdk.screens.fetch()
 }
