+++ packages/types/src/documents/app/screen.ts
@@ -29,7 +29,7 @@ export interface Screen extends Document {
   pluginAdded?: boolean
   onLoad?: EventHandler[]
   variant?: ScreenVariant
-  projectAppId?: string
+  projectAppId: string
 }
 
 export interface ScreenRoutesViewOutput extends Document {
